+++ lib/src/model/event/config/current_scene_collection_changed.g.dart
@@ -1,0 +1,19 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'current_scene_collection_changed.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+CurrentSceneCollectionChanged _$CurrentSceneCollectionChangedFromJson(
+        Map<String, dynamic> json) =>
+    CurrentSceneCollectionChanged(
+      sceneCollectionName: json['sceneCollectionName'] as String,
+    );
+
+Map<String, dynamic> _$CurrentSceneCollectionChangedToJson(
+        CurrentSceneCollectionChanged instance) =>
+    <String, dynamic>{
+      'sceneCollectionName': instance.sceneCollectionName,
+    };