+++ lib/src/model/event/config/current_scene_collection_changing.g.dart
@@ -1,0 +1,19 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'current_scene_collection_changing.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+CurrentSceneCollectionChanging _$CurrentSceneCollectionChangingFromJson(
+        Map<String, dynamic> json) =>
+    CurrentSceneCollectionChanging(
+      sceneCollectionName: json['sceneCollectionName'] as String,
+    );
+
+Map<String, dynamic> _$CurrentSceneCollectionChangingToJson(
+        CurrentSceneCollectionChanging instance) =>
+    <String, dynamic>{
+      'sceneCollectionName': instance.sceneCollectionName,
+    };